Add React Helmet
Changed how render is done on the server. Now the Document and the App
render separately, and then the app html is replaced in the Document.

This is necessary in order to make helmet work, as we need to render the
app first, so it can collect all the needed data and only then render
the whole app.

The default headers are set in Layout component.